Search functionality within search results  screen
