+++ app/styles/app.css
@@ -46,12 +46,12 @@ body {
 .books-list {
   background-color: #f9f9f9f9;
   border-radius: 8px;
-  padding: 20px;
   gap: 10px;
-  margin: 20px;
-
+  padding: 10px;
   display: grid;
   grid-template-columns: repeat(5, 1fr);
+  max-height: 100vh;
+  overflow-y: scroll;
 }
 
 .my-book {
