+++ app/styles/book-search.css
@@ -2,19 +2,21 @@
   display: grid;
   gap: 10px;
   padding: 10px;
-  display: grid;
+  padding-bottom: 75px;
   grid-template-rows: 4fr 4fr;
   grid-template-columns: 4fr 4fr;
+  max-height: 90vh;
+  overflow-y: auto;
 }
 
 .search-books-list {
-  background-color: #f9f9f9; /* Light grey background */
-  border-radius: 8px; /* Rounded corners */
-  padding: 20px; /* Spacing inside the container */
-  margin: 20px 0; /* Margin for the top and bottom */
-  max-height: 100vh; /* Maximum height of the container */
-  overflow-y: auto; /* Makes the list scrollable */
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); /* Subtle shadow for depth */
+  background-color: #f9f9f9;
+  border-radius: 8px;
+  padding: 20px;
+  margin: 20px 0;
+  max-height: 100vh; /* Adjust this value as needed */
+  overflow-y: auto; /* Use 'auto' instead of 'scroll' to only show scrollbar when necessary */
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
 }
 
 .book {
@@ -128,4 +130,37 @@
   100% {
     opacity: 0;
   }
+}
+
+.search-box {
+  display: flex;
+  justify-content: space-between;
+  padding: 10px;
+  margin-bottom: 20px; /* Spacing between the search box and the next element */
+  border-bottom: 2px solid #e1e1e1; /* Adds a subtle line to separate the search box from the content below */
+}
+
+.search-box input[type="text"] {
+  width: 80%;
+  padding: 10px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  font-size: 1em;
+  margin-left: 75px;
+}
+
+.search-box button {
+  width: 15%;
+  padding: 10px;
+  background-color: #4caf50; /* A green background for the button */
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  font-size: 1.5em;
+  margin-left: 15px;
+}
+
+.search-box button:hover {
+  background-color: #45a049; /* Darker green on hover */
 }
