+++ app/templates/index.hbs
@@ -23,7 +23,7 @@
             <p class="my-book-author">{{book.authors}}</p>
             <p class="my-book-description">{{book.description}}</p>
 
-            <!-- Add more details as needed -->
+            {{!-- Add more details as needed --}}
           </div>
         </div>
       {{else}}
